全面优化：修复安全漏洞、UX改进、性能提升
P0修复：
- 修复日记详情导航路径错误（goDetail/goReport）
- 修复积分兑换并发漏洞（SELECT FOR UPDATE锁）
- 修复情侣绑定竞态条件（双行锁定）
- 安全密钥回退改为随机生成（不再硬编码）

P1改进：
- 修复59处空catch块，全部添加错误提示toast
- 提取MOOD_CONFIG/LEVEL_POINTS/REACTION_CONFIG到共享constants.js
- 7个列表页添加下拉刷新+无限滚动分页基础设施
- Bill.amount从Float改为Numeric(10,2)避免精度问题
- 所有enum字段添加Literal类型校验（心情/反应/账单/情绪等）
- WhisperCreate.content添加max_length=2000

P2体验提升：
- 添加解除情侣绑定功能（profile页面+user store）
- 修改密码添加确认密码框
- 打卡成功/心愿完成/悄悄话发送添加震动反馈

+++ love-backend/app/models.py
@@ -3,7 +3,7 @@
 定义所有数据库表结构，使用SQLAlchemy ORM
 """
 
-from sqlalchemy import Column, Integer, String, Text, DateTime, Boolean, Float, ForeignKey, Date, UniqueConstraint
+from sqlalchemy import Column, Integer, String, Text, DateTime, Boolean, Float, Numeric, ForeignKey, Date, UniqueConstraint
 from sqlalchemy.orm import relationship
 from datetime import datetime
 from app.database import Base
@@ -313,7 +313,7 @@ class Bill(Base):
 
     id = Column(Integer, primary_key=True, index=True, autoincrement=True)
     user_id = Column(Integer, ForeignKey("users.id"), nullable=False, index=True, comment="创建人ID")
-    amount = Column(Float, nullable=False, comment="开支金额")
+    amount = Column(Numeric(10, 2), nullable=False, comment="开支金额")
     type = Column(String(20), default="other", comment="类型: food-吃饭, travel-旅行, gift-礼物, daily-日常, other-其他")
     pay_time = Column(DateTime, nullable=False, comment="开支时间")
     payer = Column(String(20), nullable=False, comment="支付人: me-我, lover-对方, aa-AA制")
